reorg + abstract `initBolt()`
