+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "prettier.tabWidth": 2,
+  "prettier.useTabs": false,
+  "prettier.printWidth": 80
+}
