+++ src/js/lib/cep/csinterface.d.ts
