+++ src/js/lib/cep/node.ts
