+++ src/js/lib/utils/aeft.ts
@@ -1,5 +1,5 @@
-import { fs, path } from "./node";
-import { csi } from "./utils";
+import { fs, path } from "../cep/node";
+import { csi } from "./bolt";
 
 const getLatestFile = (dir: string, suffix: string): string | null => {
   const getModified = (filePath: string) =>
