+++ src/js/template-react/index.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { initBolt } from "../lib/utils/bolt";
 import "../index.scss";
 import Main from "./main";
+
+initBolt();
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
