+++ src/js/template-svelte/main.ts
@@ -1,4 +1,7 @@
 import App from "./main.svelte";
+import { initBolt } from "../lib/utils/bolt";
+
+initBolt();
 
 const app = new App({
   target: document.getElementById("root") as Element,
