+++ src/js/template-vue/main.ts
@@ -1,4 +1,7 @@
 import { createApp } from "vue";
 import App from "./main.vue";
+import { initBolt } from "../lib/utils/bolt";
+
+initBolt();
 
 createApp(App).mount("#root");
