+++ tsconfig-build.json
@@ -2,7 +2,7 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "paths": {
-      "@esTypes/*": ["./src/js/lib/es-types"]
+      "@esTypes/*": ["./src/js/lib/cep/es-types"]
     }
   }
 }
